mainboard/google/fatcat: Increase RW_SECTION_A/B size by 1MB
The google/fatcat board's flash layout was modified to increase the
size of RW_SECTION_A and RW_SECTION_B by 1MB each (from 7MB to 8MB).

The RW_UNUSED region size was reduced to accommodate the increased
RW_SECTION sizes.

This change provides additional space in the RW slots to accommodate
growth in the payload (depthcharge).

TEST=Built and flashed the image. Verified that both RW_SECTION_A and
RW_SECTION_B are populated with the correct firmware components and
that the system boots successfully.

Change-Id: Ie489d53cef00ddc2dc6beef891f870c6bc0562a8
Signed-off-by: Subrata Banik <subratabanik@google.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/86202
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Jon Murphy <jpmurphy@google.com>
Reviewed-by: Jayvik Desai <jayvik@google.com>
Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>